requirements.txt
et-xmlfile==1.1.0
greenlet
numpy
openpyxl==3.1.1
pandas
playwright
pyee
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
typing_extensions==4.5.0
streamlit==1.32.0
google-generativeai==0.3.2
python-dotenv==1.0.1
